use module style that allows controlling hierarchy better
